feat: 변경된 컴포넌트 컨벤션 적용
